fix wrong data type for actions in DDPG, TD3, XDDPG, and XTD3 agents

+++ DDPG/ddpg.py
@@ -246,7 +246,7 @@ class DDPGAgent(Agent):
     states = torch.from_numpy(np.vstack([e.state for e in experiences])
                               ).float().to(device)
     actions = torch.from_numpy(np.vstack([e.action for e in experiences])
-                               ).long().to(device)
+                               ).float().to(device)
     rewards = torch.from_numpy(np.vstack([e.reward for e in experiences])
                                ).float().to(device)
     next_states = torch.from_numpy(
